requirements.txt
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.3
langchain-experimental>=0.0.40
langchain>=0.1.0
neo4j>=5.14.0
wikipedia>=1.4.0
tiktoken>=0.5.2
yfiles_jupyter_graphs>=1.9.0
pydantic>=2.5.0
openai>=1.6.0
python-dotenv>=1.0.0 